Update Notepad++ Recipe
Github is now dynamically rendering the download link in the front end of the website, and is not visible in Sources.  Because invoke-webrequest cannot render the page in IE's absence, this now looks for the "current version" link directly at notepad-plus-plus.org, then grabs the first .x64.exe file from the subsequent page, and then goes directly to the GitHub download.